small reformating to image segmentation
