+++ src/image_segmentation/utils.py
@@ -1,59 +1,6 @@
 import numbers
 import torch
-from torch.utils.data import DataLoader
-from typing import Tuple
-import os
-from src.utils import project_root
 from src.visualization_utils import show_grid_imgs
-from .weizmann_horse_dataset import WeizmannHorseDataset
-
-PATH_DATA_WEIZMANN = os.path.join(project_root(), "data", "weizmann_horse")
-PATH_SAVE_HORSE = os.path.join(project_root(), "saved_results", "weizmann_horse")
-
-
-def load_train_set_horse(path_data: str, use_cuda: bool, batch_size: int,
-                         batch_size_valid: int) -> Tuple[DataLoader, DataLoader]:
-
-    image_dir = os.path.join(path_data, 'images')
-    mask_dir = os.path.join(path_data, 'masks')
-
-    train_set = WeizmannHorseDataset(image_dir, mask_dir, subset='train',
-                                     random_mirroring=False, thirty_six_cropping=False)
-    valid_set = WeizmannHorseDataset(image_dir, mask_dir, subset='valid',
-                                     random_mirroring=False, thirty_six_cropping=False)
-
-    train_loader = DataLoader(
-        train_set,
-        batch_size=batch_size,
-        pin_memory=use_cuda
-    )
-
-    valid_loader = DataLoader(
-        valid_set,
-        batch_size=batch_size_valid,
-        pin_memory=use_cuda
-    )
-
-    print(f'Using a {len(train_loader.dataset)}/{len(valid_loader.dataset)} train/validation split')
-
-    return train_loader, valid_loader
-
-
-def load_test_set_horse(path_data: str, use_cuda: bool, batch_size: int,
-                        thirtysix_crops: bool) -> DataLoader:
-
-    image_dir = os.path.join(path_data, 'images')
-    mask_dir = os.path.join(path_data, 'masks')
-
-    test_set = WeizmannHorseDataset(image_dir, mask_dir, subset='test',
-                                    random_mirroring=False, thirty_six_cropping=thirtysix_crops)
-
-    test_loader = DataLoader(
-        test_set,
-        batch_size=batch_size,
-        pin_memory=use_cuda
-    )
-    return test_loader
 
 
 def thirty_six_crop(img, size):
